cmd: introduce template file
